Move legend components to lib
