+++ src/config/hazards/HazardLegend.tsx
@@ -1,11 +1,10 @@
 import { FC } from 'react';
 
+import { RasterLegend } from '@/lib/data-map/legend/RasterLegend';
 import { ViewLayer } from '@/lib/data-map/view-layers';
 import { formatAbbreviations } from '@/lib/react/format-abbreviations';
 
 import { HAZARD_COLOR_MAPS, HAZARDS_METADATA, HazardType } from '@/config/hazards/metadata';
-
-import { RasterLegend } from '../RasterLegend';
 
 export const HazardLegend: FC<{ viewLayer: ViewLayer<{ hazardType: HazardType }> }> = ({
   viewLayer,
