+++ src/config/natural-assets/nature-raster-layer.ts
@@ -2,11 +2,11 @@ import { Color } from 'deck.gl/typed';
 import React from 'react';
 
 import { InteractionTarget, RasterTarget } from '@/lib/data-map/interactions/types';
+import { RasterColorMap, RasterLegend } from '@/lib/data-map/legend/RasterLegend';
 import { ViewLayer } from '@/lib/data-map/view-layers';
 import { rasterTileLayer } from '@/lib/deck/layers/raster-tile-layer';
 import { toLabelLookup } from '@/lib/helpers';
 
-import { RasterColorMap, RasterLegend } from '@/map/legend/RasterLegend';
 import { RasterHoverDescription } from '@/map/tooltip/RasterHoverDescription';
 
 import { SOURCES } from '../sources';
