+++ src/config/travel-time/travel-time-layer.ts
@@ -2,10 +2,10 @@ import { capitalize } from 'lodash';
 import React from 'react';
 
 import { InteractionTarget, RasterTarget } from '@/lib/data-map/interactions/types';
+import { RasterColorMap, RasterLegend } from '@/lib/data-map/legend/RasterLegend';
 import { rasterTileLayer } from '@/lib/deck/layers/raster-tile-layer';
 import { makeValueFormat, nullFormat } from '@/lib/formats';
 
-import { RasterColorMap, RasterLegend } from '@/map/legend/RasterLegend';
 import { RasterHoverDescription } from '@/map/tooltip/RasterHoverDescription';
 
 import { SOURCES } from '../sources';
